Refactor app_insight layout.
