+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/cameras/api/views.py
@@ -6,12 +6,8 @@ from __future__ import absolute_import, unicode_literals
 
 import logging
 
-from django.http import JsonResponse
 from filters.mixins import FiltersMixin
 from rest_framework import filters, viewsets
-from rest_framework.decorators import api_view
-
-from configs.app_insight import APP_INSIGHT_INST_KEY
 
 from ..models import Camera
 from .serializers import CameraSerializer
@@ -34,9 +30,3 @@ class CameraViewSet(FiltersMixin, viewsets.ModelViewSet):
     filter_mappings = {
         "is_demo": "is_demo",
     }
-
-
-@api_view()
-def instrumentation_key(request):
-    """App Insight Instrument Key"""
-    return JsonResponse({"status": "ok", "key": APP_INSIGHT_INST_KEY})
